# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blue Media social platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /users:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users
      parameters:
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get user stats
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [follows]
      summary: Follow a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
    delete:
      operationId: unfollowUser
      tags: [follows]
      summary: Unfollow a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /users/{id}/report:
    post:
      operationId: reportUser
      tags: [reports]
      summary: Report a user profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /friends:
    get:
      operationId: listFriends
      tags: [friends]
      summary: List friends
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
      responses:
        "200":
          description: Friends list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: Get home feed posts
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Get or create a direct conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]
    RegisterInput:
      type: object
      required: [email, name, pin]
      properties:
        email:
          type: string
        name:
          type: string
        pin:
          type: string
          minLength: 4
          maxLength: 4
    LoginInput:
      type: object
      required: [email, pin]
      properties:
        email:
          type: string
        pin:
          type: string
...
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        profilePicture:
          type: ["string", "null"]
        coverPicture:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        privacy:
          type: string
          enum: [public, friends_of_friends, private]
        isAdmin:
          type: boolean
        isBlueAI:
          type: boolean
        blueBadge:
          type: boolean
        blueBadgeClaimedAt:
          type: ["string", "null"]
          format: date-time
        restricted:
          type: boolean
        banned:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        privacy:
          type: string
          enum: [public, friends_of_friends, private]
    UserStats:
      type: object
      required: [friendCount, postCount, followerCount, followingCount]
      properties:
        friendCount:
          type: integer
        postCount:
          type: integer
        followerCount:
          type: integer
        followingCount:
          type: integer
...
    Post:
      type: object
      required: [id, userId, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        bgColor:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        liveStreamId:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, friends_of_friends, private]
        author:
          $ref: "#/components/schemas/User"
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/PostReaction"
        commentCount:
          type: integer
        myReaction:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        bgColor:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        liveStreamId:
          type: ["string", "null"]
        feeling:
          type: ["string", "null"]
        activity:
          type: ["string", "null"]
        locationTag:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, friends_of_friends, private]
    PostReaction:
      type: object
      required: [type, count]
      properties:
        type:
          type: string
        count:
          type: integer
...
    ReportInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum:
            - sexual_content
            - harassment
            - hate_speech
            - violence
            - spam
...
    Conversation:
      type: object
      required: [id, type, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [direct, group]
        name:
          type: ["string", "null"]
        pictureUrl:
          type: ["string", "null"]
        backgroundTheme:
          type: ["string", "null"]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ConversationInput:
      type: object
      required: [participantId]
      properties:
        participantId:
          type: integer
...
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        sender:
          $ref: "#/components/schemas/User"
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/MessageReaction"
        seenBy:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
...
    MessageReaction:
      type: object
      required: [id, messageId, userId, emoji]
      properties:
        id:
          type: integer
        messageId:
          type: integer
        userId:
          type: integer
        emoji:
          type: string
        user:
          $ref: "#/components/schemas/User"
...